+++ backend/package.json
@@ -8,6 +8,7 @@
     "cookie-parser": "^1.4.5",
     "cors": "^2.8.5",
     "express": "^4.17.1",
+    "ip": "^1.1.5",
     "jsonwebtoken": "^8.5.1",
     "koa": "^2.13.0",
     "koa-bodyparser": "^4.3.0",
@@ -19,9 +20,10 @@
     "koa-send": "^5.0.1",
     "koa-static": "^5.0.0",
     "mongoose": "^5.11.4",
-    "mysql": "^2.18.1"
+    "mysql": "^2.18.1",
+    "public-ip": "^4.0.3"
   },
   "scripts": {
-    "start": "npx nodemon index.js"
+    "start": "npx nodemon ./src/index.js"
   }
 }
