+++ backend/src/middleware/auth.js
