+++ frontend/dockerfile
@@ -1,5 +1,5 @@
 FROM node:13
 COPY . /app
 WORKDIR /app
-RUN npm i
 RUN yarn
+CMD [ "yarn","run","start" ]
